パッケージをドメイン名に変更
