+++ src/main/java/com/herokuapp/issue_quest/IssueQuestApplication.java
@@ -1,0 +1,12 @@
+package com.herokuapp.issue_quest;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class IssueQuestApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(IssueQuestApplication.class, args);
+	}
+}
