+++ src/main/java/com/herokuapp/issue_quest/app/QuestForm.java
@@ -1,0 +1,22 @@
+package com.herokuapp.issue_quest.app;
+
+import lombok.Getter;
+import lombok.Setter;
+
+/**
+ * クエスト作成のためのフォーム
+ * @author k4h4shi
+ *
+ */
+@Getter
+@Setter
+public class QuestForm {
+  /**
+   * タイトル
+   */
+  private String title;
+  /**
+   * 説明
+   */
+  private String description;
+}
