+++ src/main/java/com/herokuapp/issue_quest/config/SecurityConfig.java
@@ -1,0 +1,49 @@
+package com.herokuapp.issue_quest.config;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+import com.herokuapp.issue_quest.auth.service.IssueQuestUserDetailsService;
+
+/**
+ * spring-security設定クラス
+ * @author obscure12
+ *
+ */
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig extends WebSecurityConfigurerAdapter{
+
+	@Autowired
+	@Qualifier("issueQuestUserDetailsService")
+	private IssueQuestUserDetailsService issueQuestUserDetailsService;
+
+	@Override
+	protected void configure(HttpSecurity http) throws Exception {
+		http
+			.authorizeRequests()
+				.anyRequest().fullyAuthenticated()
+				.and()
+			.formLogin()
+				.defaultSuccessUrl("/index.html")
+				.and()
+			.logout()
+				.logoutUrl("/logout")
+				.logoutSuccessUrl("/")
+				.and()
+			.csrf()
+				.disable();
+	}
+
+	@Bean
+	@Override
+	public AuthenticationManager authenticationManagerBean() throws Exception {
+		return super.authenticationManagerBean();
+	}
+}
