+++ src/main/java/com/herokuapp/issue_quest/domain/model/QuestStatus.java
@@ -1,0 +1,20 @@
+package com.herokuapp.issue_quest.domain.model;
+/**
+ * クエストのステータスのenum
+ * @author k4h4shi
+ *
+ */
+public enum QuestStatus {
+  /**
+   * 未受注
+   */
+  UNDONE,
+  /**
+   * 受注中
+   */
+  DOING,
+  /**
+   * 完了
+   */
+  DONE;
+}
