--- src/main/java/com/k4h4shi/issueQuest/IssueQuestApplication.java
@@ -1,12 +1,0 @@
-package com.k4h4shi.issueQuest;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class IssueQuestApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(IssueQuestApplication.class, args);
-	}
-}
