--- src/main/java/com/k4h4shi/issueQuest/domain/model/QuestStatus.java
@@ -1,20 +1,0 @@
-package com.k4h4shi.issueQuest.domain.model;
-/**
- * クエストのステータスのenum
- * @author k4h4shi
- *
- */
-public enum QuestStatus {
-  /**
-   * 未受注
-   */
-  UNDONE,
-  /**
-   * 受注中
-   */
-  DOING,
-  /**
-   * 完了
-   */
-  DONE;
-}
