+++ src/test/java/com/herokuapp/issue_quest/IssueQuestApplicationTests.java
@@ -1,0 +1,16 @@
+package com.herokuapp.issue_quest;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class IssueQuestApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
