--- src/test/java/com/k4h4shi/issueQuest/IssueQuestApplicationTests.java
@@ -1,16 +1,0 @@
-package com.k4h4shi.issueQuest;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class IssueQuestApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
